Desarrollo_ServidorBasicoDjango/requirements.txt
pandas
Django
requests
djangorestframework